Moved query and handler to same file.
