--- BingeBuddyNg/BingeBuddyNg.Services/Activity/Querys/GetActivitysForMapQuery.cs
@@ -1,17 +1,0 @@
-﻿using MediatR;
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace BingeBuddyNg.Services.Activity.Querys
-{
-    public class GetActivitysForMapQuery : IRequest<List<ActivityDTO>>
-    {
-        public string UserId { get; }
-
-        public GetActivitysForMapQuery(string userId)
-        {
-            this.UserId = userId ?? throw new ArgumentNullException(nameof(userId));
-        }
-    }
-}
