+++ BingeBuddyNg/BingeBuddyNg.Spa/src/app/activity/components/activity-feed/activity-feed.component.ts
@@ -74,7 +74,6 @@ export class ActivityFeedComponent implements OnInit, OnDestroy {
     public locationService: LocationService,
     private drinkActivityService: DrinkActivityService,
     private drinkService: DrinkRetrieverService,
-    private changeRef: ChangeDetectorRef,
     private snackBar: MatSnackBar,
     private translateService: TranslocoService,
     private dialog: MatDialog,
@@ -356,7 +355,6 @@ export class ActivityFeedComponent implements OnInit, OnDestroy {
     this.isBusyAdding = false;
     this.isBusyUploading = true;
     this.currentProgress = progress;
-    this.changeRef.detectChanges();
   }
 
   onActionsOpenChange(open: boolean): void {
